requirements.txt
Django==1.7.9
djangorestframework==3.3.0
psycopg2==2.6.1
celery==3.1.24
django-celery==3.1.17
kombu==3.0.37
xmltodict==0.10.2